# Avoid recreating entry_point object.