@jwodder 2022-01-16
Thank you for the analysis and patch!  I will release a new version with this momentarily.