@jwodder 2022-01-16
versioningit 0.3.2 released.